Remove unused image ops, simplify size header.
Encoder is 9.4 kLOC, benchmark results:

```
Encoding                           kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
---------------------------------------------------------------------------------------------------------------------------------
custom:jxl:cjxl_tiny:djxl:-d1        13270  2550029    1.5372677   3.041   4.479   1.75526035   0.65428949  1.005818062815      0
custom:jxl:cjxl_tiny:djxl:-d4        13270   918875    0.5539375   3.269   3.607   5.10233402   1.68180702  0.931616062027      0
```
