+++ encoder/chroma_from_luma.h
@@ -15,7 +15,6 @@
 
 #include "encoder/common.h"
 #include "encoder/image.h"
-#include "encoder/image_ops.h"
 
 namespace jxl {
