+++ encoder/dct_util.h
@@ -14,7 +14,6 @@
 #include "encoder/base/status.h"
 #include "encoder/common.h"
 #include "encoder/image.h"
-#include "encoder/image_ops.h"
 
 namespace jxl {
